Simplifies handling of entering app from obscure passcode
